+++ .storybook/manager.js
@@ -1,0 +1,9 @@
+import { addons } from '@storybook/addons';
+import wingsTheme from './theme/theme';
+import '@storybook/addon-knobs/register';
+import '@storybook/addon-viewport/register';
+import '@storybook/addon-contexts/register';
+
+addons.setConfig({
+  theme: wingsTheme,
+});