+++ packages/components/src/components/Icon/icons/index.js
@@ -1,0 +1,69 @@
+import ArrowDown from './ArrowDown';
+import AudioOn from './AudioOn';
+import AudioOff from './AudioOff';
+import Calendar from './Calendar';
+import Check from './Check';
+import ChevronDown from './ChevronDown';
+import ChevronLeft from './ChevronLeft';
+import ChevronRight from './ChevronRight';
+import ChevronUp from './ChevronUp';
+import Email from './Email';
+import Eye from './Eye';
+import Instagram from './Instagram';
+import Facebook from './Facebook';
+import Github from './Github';
+import Globe from './Globe';
+import Linkedin from './Linkedin';
+import Message from './Message';
+import Pause from './Pause';
+import Pin from './Pin';
+import Pinterest from './Pinterest';
+import Play from './Play';
+import Quote from './Quote';
+import Reddit from './Reddit';
+import Share from './Share';
+import Skype from './Skype';
+import Spotify from './Spotify';
+import Twitch from './Twitch';
+import Twitter from './Twitter';
+import User from './User';
+import Vimeo from './Vimeo';
+import RSS from './RSS';
+import Whatsapp from './Whatsapp';
+import Youtube from './Youtube';
+
+export default {
+  ArrowDown,
+  AudioOn,
+  AudioOff,
+  Calendar,
+  Check,
+  ChevronDown,
+  ChevronLeft,
+  ChevronRight,
+  ChevronUp,
+  Email,
+  Eye,
+  Instagram,
+  Facebook,
+  Github,
+  Globe,
+  Linkedin,
+  Message,
+  Pause,
+  Pin,
+  Pinterest,
+  Play,
+  Quote,
+  Reddit,
+  Share,
+  Skype,
+  Spotify,
+  Twitch,
+  Twitter,
+  User,
+  Vimeo,
+  RSS,
+  Whatsapp,
+  Youtube,
+};