+++ packages/components/src/lib/utils/randomString.js
@@ -1,5 +1,0 @@
-export default (length = 10) =>
-  Math.random()
-    .toString(length)
-    .replace(/[^a-zA-Z0-9]+/g, '')
-    .substr(0, 5);