+++ packages/react/src/cards/CollectionCard.js
@@ -1,10 +1,8 @@
 import React, { Component } from 'react';
 
-import { FlexGrid, Icons, wide } from '@wingscms/components';
+import { FlexGrid, Icon, wide } from '@wingscms/components';
 import styled from '../lib/styled';
 import createCard from '../createCard';
-
-const { Facebook, Globe, Instagram, Linkedin, Twitter } = Icons;
 
 const Wide = styled.div`
   ${wide}
@@ -145,21 +143,21 @@ class CollectionCardView extends Component {
               {x.profiles.website ? (
                 <a href={x.profiles.website.url} target="_blank" rel="noopener noreferrer">
                   <ProfileImage>
-                    <Globe />
+                    <Icon icon="globe" />
                   </ProfileImage>
                 </a>
               ) : null}
               {x.profiles.facebook ? (
                 <a href={x.profiles.facebook.url} target="_blank" rel="noopener noreferrer">
                   <ProfileImage>
-                    <Facebook />
+                    <Icon icon="facebook" />
                   </ProfileImage>
                 </a>
               ) : null}
               {x.profiles.linkedIn ? (
                 <a href={x.profiles.linkedIn.url} target="_blank" rel="noopener noreferrer">
                   <ProfileImage>
-                    <Linkedin />
+                    <Icon icon="linked-in" />
                   </ProfileImage>
                 </a>
               ) : null}
@@ -170,7 +168,7 @@ class CollectionCardView extends Component {
                   rel="noopener noreferrer"
                 >
                   <ProfileImage>
-                    <Instagram />
+                    <Icon icon="instagram" />
                   </ProfileImage>
                 </a>
               ) : null}
@@ -181,7 +179,7 @@ class CollectionCardView extends Component {
                   rel="noopener noreferrer"
                 >
                   <ProfileImage>
-                    <Twitter />
+                    <Icon icon="twitter" />
                   </ProfileImage>
                 </a>
               ) : null}