+++ packages/react/src/components/MobiledocRenderer.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
+import filterInvalidDOMProps from 'filter-invalid-dom-props';
 import Renderer03 from '@dailybeast/mobiledoc-react-renderer/dist/renderers/0-3';
 import Renderer from '@dailybeast/mobiledoc-react-renderer';
 import { Link } from '@wingscms/components';
@@ -52,7 +53,11 @@ class MobiledocRenderer extends Component {
   render() {
     const { content, cards: _cards, unknownCardHandler, cardProps: _, ...props } = this.props;
     const renderer = this.createRenderer();
-    return <div {...props}>{!content ? null : renderer.render(JSON.parse(content))}</div>;
+    return (
+      <div {...filterInvalidDOMProps(props)}>
+        {!content ? null : renderer.render(JSON.parse(content))}
+      </div>
+    );
   }
 }