+++ packages/react/src/components/Node.stories.js
@@ -1,4 +1,0 @@
-import React from 'react';
-// import Node from './Node';
-
-export default () => 'asdasd'; //<Node node={{}} />;